Finish refactoring tbl_*.lib.php
