+++ libraries/central_columns.lib.php
@@ -9,6 +9,8 @@ if (! defined('PHPMYADMIN')) {
     exit;
 }
 
+require_once './libraries/Template.class.php';
+
 /**
  * Defines the central_columns parameters for the current user
  *
@@ -93,8 +95,8 @@ function PMA_getCentralColumnsCount($db)
     $GLOBALS['dbi']->selectDb($pmadb, $GLOBALS['controllink']);
     $central_list_table = $cfgCentralColumns['table'];
     $query = 'SELECT count(db_name) FROM ' .
-               PMA_Util::backquote($central_list_table) . ' '
-            . 'WHERE db_name = \'' . $db . '\';';
+        PMA_Util::backquote($central_list_table) . ' '
+        . 'WHERE db_name = \'' . $db . '\';';
     $res = $GLOBALS['dbi']->fetchResult(
         $query, null, null, $GLOBALS['controllink']
     );
@@ -189,15 +191,15 @@ function PMA_getInsertQuery($column, $def, $db, $central_list_table)
     $extra = isset($def['Extra'])?$def['Extra']:"";
     $default = isset($def['Default'])?$def['Default']:"";
     $insQuery = 'INSERT INTO '
-    . PMA_Util::backquote($central_list_table) . ' '
-    . 'VALUES ( \'' . PMA_Util::sqlAddSlashes($db) . '\' ,'
-    . '\'' . PMA_Util::sqlAddSlashes($column) . '\',\''
-    . PMA_Util::sqlAddSlashes($type) . '\','
-    . '\'' . PMA_Util::sqlAddSlashes($length) . '\',\''
-    . PMA_Util::sqlAddSlashes($collation) . '\','
-    . '\'' . PMA_Util::sqlAddSlashes($isNull) . '\','
-    . '\'' . implode(',', array($extra, $attribute))
-    . '\',\'' . PMA_Util::sqlAddSlashes($default) . '\');';
+        . PMA_Util::backquote($central_list_table) . ' '
+        . 'VALUES ( \'' . PMA_Util::sqlAddSlashes($db) . '\' ,'
+        . '\'' . PMA_Util::sqlAddSlashes($column) . '\',\''
+        . PMA_Util::sqlAddSlashes($type) . '\','
+        . '\'' . PMA_Util::sqlAddSlashes($length) . '\',\''
+        . PMA_Util::sqlAddSlashes($collation) . '\','
+        . '\'' . PMA_Util::sqlAddSlashes($isNull) . '\','
+        . '\'' . implode(',', array($extra, $attribute))
+        . '\',\'' . PMA_Util::sqlAddSlashes($default) . '\');';
     return $insQuery;
 }
 
@@ -259,7 +261,7 @@ function PMA_syncUniqueColumns($field_select, $isTable=true, $table=null)
         foreach ($field_select as $column) {
             $cols .= "'" . PMA_Util::sqlAddSlashes($column) . "',";
         }
-         $has_list = PMA_findExistingColNames($db, trim($cols, ','));
+        $has_list = PMA_findExistingColNames($db, trim($cols, ','));
         foreach ($field_select as $column) {
             if (!in_array($column, $has_list)) {
                 $has_list[] = $column;
@@ -365,21 +367,21 @@ function PMA_deleteColumnsFromList($field_select, $isTable=true)
             }
         }
     }
-    if (! empty($colNotExist)) {
-            $colNotExist = implode(",", array_unique($colNotExist));
-            $message = PMA_Message::notice(
-                sprintf(
-                    __(
-                        'Couldn\'t remove Column(s) %1$s '
-                        . 'as they don\'t exist in central columns list!'
-                    ), htmlspecialchars($colNotExist)
-                )
-            );
+    if (!empty($colNotExist)) {
+        $colNotExist = implode(",", array_unique($colNotExist));
+        $message = PMA_Message::notice(
+            sprintf(
+                __(
+                    'Couldn\'t remove Column(s) %1$s '
+                    . 'as they don\'t exist in central columns list!'
+                ), htmlspecialchars($colNotExist)
+            )
+        );
     }
     $GLOBALS['dbi']->selectDb($pmadb, $GLOBALS['controllink']);
 
     $query = 'DELETE FROM ' . PMA_Util::backquote($central_list_table) . ' '
-                . 'WHERE db_name = \'' . $db . '\' AND col_name IN (' . $cols . ');';
+        . 'WHERE db_name = \'' . $db . '\' AND col_name IN (' . $cols . ');';
 
     if (!$GLOBALS['dbi']->tryQuery($query, $GLOBALS['controllink'])) {
         $message = PMA_Message::error(__('Could not remove columns!'));
@@ -433,12 +435,12 @@ function PMA_makeConsistentWithList($db, $selected_tables)
                 if ($column['col_default']) {
                     if ($column['col_default'] != 'CURRENT_TIMESTAMP') {
                         $query .= ' DEFAULT \'' . PMA_Util::sqlAddSlashes(
-                            $column['col_default']
-                        ) . '\'';
+                                $column['col_default']
+                            ) . '\'';
                     } else {
                         $query .= ' DEFAULT ' . PMA_Util::sqlAddSlashes(
-                            $column['col_default']
-                        );
+                                $column['col_default']
+                            );
                     }
                 }
                 $query .= ',';
@@ -512,7 +514,7 @@ function PMA_getCentralColumnsFromTable($db, $table, $allFields=false)
  * @return true|PMA_Message
  */
 function PMA_updateOneColumn($db, $orig_col_name, $col_name, $col_type,
-    $col_attribute,$col_length, $col_isNull, $collation, $col_extra, $col_default
+                             $col_attribute,$col_length, $col_isNull, $collation, $col_extra, $col_default
 ) {
     $cfgCentralColumns = PMA_centralColumnsGetParams();
     if (empty($cfgCentralColumns)) {
@@ -534,24 +536,24 @@ function PMA_updateOneColumn($db, $orig_col_name, $col_name, $col_type,
         $query = PMA_getInsertQuery($col_name, $def, $db, $centralTable);
     } else {
         $query = 'UPDATE ' . PMA_Util::backquote($centralTable)
-                . ' SET col_type = \'' . PMA_Util::sqlAddSlashes($col_type) . '\''
-                . ', col_name = \'' . PMA_Util::sqlAddSlashes($col_name) . '\''
-                . ', col_length = \'' . PMA_Util::sqlAddSlashes($col_length) . '\''
-                . ', col_isNull = ' . $col_isNull
-                . ', col_collation = \'' . PMA_Util::sqlAddSlashes($collation) . '\''
-                . ', col_extra = \''
-                . implode(',', array($col_extra, $col_attribute)) . '\''
-                . ', col_default = \'' . PMA_Util::sqlAddSlashes($col_default) . '\''
-                . ' WHERE db_name = \'' . PMA_Util::sqlAddSlashes($db) . '\' '
-                . 'AND col_name = \'' . PMA_Util::sqlAddSlashes($orig_col_name)
-                . '\'';
+            . ' SET col_type = \'' . PMA_Util::sqlAddSlashes($col_type) . '\''
+            . ', col_name = \'' . PMA_Util::sqlAddSlashes($col_name) . '\''
+            . ', col_length = \'' . PMA_Util::sqlAddSlashes($col_length) . '\''
+            . ', col_isNull = ' . $col_isNull
+            . ', col_collation = \'' . PMA_Util::sqlAddSlashes($collation) . '\''
+            . ', col_extra = \''
+            . implode(',', array($col_extra, $col_attribute)) . '\''
+            . ', col_default = \'' . PMA_Util::sqlAddSlashes($col_default) . '\''
+            . ' WHERE db_name = \'' . PMA_Util::sqlAddSlashes($db) . '\' '
+            . 'AND col_name = \'' . PMA_Util::sqlAddSlashes($orig_col_name)
+            . '\'';
     }
     if (!$GLOBALS['dbi']->tryQuery($query, $GLOBALS['controllink'])) {
         return PMA_Message::error(
             $GLOBALS['dbi']->getError($GLOBALS['controllink'])
         );
     }
-     return true;
+    return true;
 }
 
 /**
@@ -592,7 +594,7 @@ function PMA_updateMultipleColumn()
             return $message;
         }
     }
-     return true;
+    return true;
 }
 
 /**
@@ -862,23 +864,44 @@ function PMA_getHTMLforCentralColumnsTableRow($row, $odd_row, $row_num, $db)
         . '<span>' . htmlspecialchars($row['col_name']) . '</span>'
         . '<input name="orig_col_name" type="hidden" '
         . 'value="' . htmlspecialchars($row['col_name']) . '">'
-        . PMA_getHtmlForColumnName(
-            $row_num, 0, 0, array('Field'=>$row['col_name']),
-            array('central_columnswork'=>false)
-        )
+        . PMA\Template::get('columns_definitions/column_name')
+            ->render(array(
+                'columnNumber' => $row_num,
+                'ci' => 0,
+                'ci_offset' => 0,
+                'columnMeta' => array(
+                    'Field'=>$row['col_name']
+                ),
+                'cfgRelation' => array(
+                    'central_columnswork' => false
+                )
+            ))
         . '</td>';
     $tableHtml .=
         '<td name = "col_type" class="nowrap"><span>'
         . htmlspecialchars($row['col_type']) . '</span>'
-        . PMA_getHtmlForColumnType(
-            $row_num, 1, 0, /*overload*/mb_strtoupper($row['col_type']), array()
-        )
+        . PMA\Template::get('columns_definitions/column_type')
+            ->render(array(
+                'columnNumber' => $row_num,
+                'ci' => 1,
+                'ci_offset' => 0,
+                'type_upper' => /*overload*/mb_strtoupper($row['col_type']),
+                'columnMeta' => array()
+            ))
         . '</td>';
     $tableHtml .=
         '<td class="nowrap" name="col_length">'
         . '<span>' . ($row['col_length']?htmlspecialchars($row['col_length']):"")
         . '</span>'
-        . PMA_getHtmlForColumnLength($row_num, 2, 0, 8, $row['col_length'])
+        . PMA\Template::get('columns_definitions/column_length')->render(
+            array(
+                'columnNumber' => $row_num,
+                'ci' => 2,
+                'ci_offset' => 0,
+                'length_values_input_size' => 8,
+                'length_to_display' => $row['col_length']
+            )
+        )
         . '</td>';
 
     $meta = array();
@@ -896,42 +919,68 @@ function PMA_getHTMLforCentralColumnsTableRow($row, $odd_row, $row_num, $db)
     }
     $tableHtml .=
         '<td class="nowrap" name="col_default"><span>' . (isset($row['col_default'])
-        ? htmlspecialchars($row['col_default']) : 'None')
+            ? htmlspecialchars($row['col_default']) : 'None')
         . '</span>'
-        . PMA_getHtmlForColumnDefault(
-            $row_num, 3, 0, /*overload*/mb_strtoupper($row['col_type']), '', $meta
-        )
+        . PMA\Template::get('columns_definitions/column_default')
+            ->render(array(
+                'columnNumber' => $row_num,
+                'ci' => 3,
+                'ci_offset' => 0,
+                'type_upper' => /*overload*/mb_strtoupper($row['col_type']),
+                'columnMeta' => $meta
+            ))
         . '</td>';
 
     $tableHtml .=
         '<td name="collation" class="nowrap">'
         . '<span>' . htmlspecialchars($row['col_collation']) . '</span>'
-        . PMA_getHtmlForColumnCollation(
-            $row_num, 4, 0, array('Collation'=>$row['col_collation'])
+        . PMA_generateCharsetDropdownBox(
+            PMA_CSDROPDOWN_COLLATION, 'field_collation[' . $row_num . ']',
+            'field_' . $row_num . '_4', $row['col_collation'], false
         )
         . '</td>';
     $tableHtml .=
         '<td class="nowrap" name="col_attribute">'
         . '<span>' .
         ($row['col_attribute']
-        ? htmlspecialchars($row['col_attribute']) : "" )
+            ? htmlspecialchars($row['col_attribute']) : "" )
         . '</span>'
-        . PMA_getHtmlForColumnAttribute(
-            $row_num, 5, 0, array(), $row['col_attribute'], false, null
-        )
+        . PMA\Template::get('columns_definitions/column_attribute')
+            ->render(array(
+                'columnNumber' => $row_num,
+                'ci' => 5,
+                'ci_offset' => 0,
+                'extracted_columnspec' => array(),
+                'columnMeta' => $row['col_attribute'],
+                'submit_attribute' => false,
+                'analyzed_sql' => null
+            ))
         . '</td>';
     $tableHtml .=
         '<td class="nowrap" name="col_isNull">'
         . '<span>' . ($row['col_isNull'] ? __('Yes') : __('No'))
         . '</span>'
-        . PMA_getHtmlForColumnNull($row_num, 6, 0, array('Null'=>$row['col_isNull']))
+        . PMA\Template::get('columns_definitions/column_null')
+            ->render(array(
+                'columnNumber' => $row_num,
+                'ci' => 6,
+                'ci_offset' => 0,
+                'columnMeta' => array(
+                    'Null' => $row['col_isNull']
+                )
+            ))
         . '</td>';
 
     $tableHtml .=
         '<td class="nowrap" name="col_extra"><span>'
         . htmlspecialchars($row['col_extra']) . '</span>'
-        . PMA_getHtmlForColumnExtra(
-            $row_num, 7, 0, array('Extra'=>$row['col_extra'])
+        . PMA\Template::get('columns_definitions/column_extra')->render(
+            array(
+                'columnNumber' => $row_num,
+                'ci' => 7,
+                'ci_offset' => 0,
+                'columnMeta' => array('Extra'=>$row['col_extra'])
+            )
         )
         . '</td>';
 
@@ -956,20 +1005,41 @@ function PMA_getHTMLforCentralColumnsEditTableRow($row, $odd_row, $row_num)
         . '<input name="orig_col_name[' . $row_num . ']" type="hidden" '
         . 'value="' . htmlspecialchars($row['col_name']) . '">'
         . '<td name="col_name" class="nowrap">'
-        . PMA_getHtmlForColumnName(
-            $row_num, 0, 0, array('Field'=>$row['col_name']),
-            array('central_columnswork'=>false)
-        )
+        . PMA\Template::get('columns_definitions/column_name')
+            ->render(array(
+                'columnNumber' => $row_num,
+                'ci' => 0,
+                'ci_offset' => 0,
+                'columnMeta' => array(
+                    'Field' => $row['col_name']
+                ),
+                'cfgRelation' => array(
+                    'central_columnswork' => false
+                )
+            ))
         . '</td>';
     $tableHtml .=
         '<td name = "col_type" class="nowrap">'
-        . PMA_getHtmlForColumnType(
-            $row_num, 1, 0, /*overload*/mb_strtoupper($row['col_type']), array()
-        )
+        . PMA\Template::get('columns_definitions/column_type')
+            ->render(array(
+                'columnNumber' => $row_num,
+                'ci' => 1,
+                'ci_offset' => 0,
+                'type_upper' => /*overload*/mb_strtoupper($row['col_type']),
+                'columnMeta' => array()
+            ))
         . '</td>';
     $tableHtml .=
         '<td class="nowrap" name="col_length">'
-        . PMA_getHtmlForColumnLength($row_num, 2, 0, 8, $row['col_length'])
+        . PMA\Template::get('columns_definitions/column_length')->render(
+            array(
+                'columnNumber' => $row_num,
+                'ci' => 2,
+                'ci_offset' => 0,
+                'length_values_input_size' => 8,
+                'length_to_display' => $row['col_length']
+            )
+        )
         . '</td>';
     $meta = array();
     if (!isset($row['col_default']) || $row['col_default'] == '') {
@@ -986,32 +1056,59 @@ function PMA_getHTMLforCentralColumnsEditTableRow($row, $odd_row, $row_num)
     }
     $tableHtml .=
         '<td class="nowrap" name="col_default">'
-        . PMA_getHtmlForColumnDefault(
-            $row_num, 3, 0, /*overload*/mb_strtoupper($row['col_default']), '', $meta
-        )
+        . PMA\Template::get('columns_definitions/column_default')
+            ->render(array(
+                'columnNumber' => $row_num,
+                'ci' => 3,
+                'ci_offset' => 0,
+                'type_upper' => /*overload*/mb_strtoupper($row['col_default']),
+                'columnMeta' => $meta
+            ))
         . '</td>';
     $tableHtml .=
         '<td name="collation" class="nowrap">'
-        . PMA_getHtmlForColumnCollation(
-            $row_num, 4, 0, array('Collation'=>$row['col_collation'])
+        . PMA_generateCharsetDropdownBox(
+            PMA_CSDROPDOWN_COLLATION, 'field_collation[' . $row_num . ']',
+            'field_' . $row_num . '_4', $row['col_collation'], false
         )
         . '</td>';
     $tableHtml .=
         '<td class="nowrap" name="col_attribute">'
-        . PMA_getHtmlForColumnAttribute(
-            $row_num, 5, 0, array("attribute"=>$row['col_attribute']),
-            array(), false, null
-        )
+        . PMA\Template::get('columns_definitions/column_attribute')
+            ->render(array(
+                'columnNumber' => $row_num,
+                'ci' => 5,
+                'ci_offset' => 0,
+                'extracted_columnspec' => array(
+                    'attribute' => $row['col_attribute']
+                ),
+                'columnMeta' => array(),
+                'submit_attribute' => false,
+                'analyzed_sql' => null
+            ))
         . '</td>';
     $tableHtml .=
         '<td class="nowrap" name="col_isNull">'
-        . PMA_getHtmlForColumnNull($row_num, 6, 0, array('Null'=>$row['col_isNull']))
+        . PMA\Template::get('columns_definitions/column_null')
+            ->render(array(
+                'columnNumber' => $row_num,
+                'ci' => 6,
+                'ci_offset' => 0,
+                'columnMeta' => array(
+                    'Null' => $row['col_isNull']
+                )
+            ))
         . '</td>';
 
     $tableHtml .=
         '<td class="nowrap" name="col_extra">'
-        . PMA_getHtmlForColumnExtra(
-            $row_num, 7, 0, array('Extra' => $row['col_extra'])
+        . PMA\Template::get('columns_definitions/column_extra')->render(
+            array(
+                'columnNumber' => $row_num,
+                'ci' => 7,
+                'ci_offset' => 0,
+                'columnMeta' => array('Extra' => $row['col_extra'])
+            )
         )
         . '</td>';
     $tableHtml .= '</tr>';
@@ -1036,7 +1133,7 @@ function PMA_getCentralColumnsListRaw($db, $table)
     $centralTable = $cfgCentralColumns['table'];
     if (empty($table) || $table == '') {
         $query = 'SELECT * FROM ' . PMA_Util::backquote($centralTable) . ' '
-                . 'WHERE db_name = \'' . $db . '\';';
+            . 'WHERE db_name = \'' . $db . '\';';
     } else {
         $GLOBALS['dbi']->selectDb($db, $GLOBALS['userlink']);
         $columns = (array) $GLOBALS['dbi']->getColumnNames(
@@ -1157,32 +1254,84 @@ function PMA_getHTMLforAddNewColumn($db)
     $addNewColumn .= '<tr>'
         . '<td></td>'
         . '<td name="col_name" class="nowrap">'
-        .  PMA_getHtmlForColumnName(
-            0, 0, 0, array(), array('central_columnswork'=>false)
-        )
+        . PMA\Template::get('columns_definitions/column_name')
+            ->render(array(
+                'columnNumber' => 0,
+                'ci' => 0,
+                'ci_offset' => 0,
+                'columnMeta' => array(),
+                'cfgRelation' => array(
+                    'central_columnswork' => false
+                )
+            ))
         . '</td>'
         . '<td name = "col_type" class="nowrap">'
-        .  PMA_getHtmlForColumnType(0, 1, 0, '', array())
+        . PMA\Template::get('columns_definitions/column_type')
+            ->render(array(
+                'columnNumber' => 0,
+                'ci' => 1,
+                'ci_offset' => 0,
+                'type_upper' => '',
+                'columnMeta' => array()
+            ))
         . '</td>'
         . '<td class="nowrap" name="col_length">'
-        . PMA_getHtmlForColumnLength(0, 2, 0, 8, '')
+        . PMA\Template::get('columns_definitions/column_length')->render(
+            array(
+                'columnNumber' => 0,
+                'ci' => 2,
+                'ci_offset' => 0,
+                'length_values_input_size' => 8,
+                'length_to_display' => ''
+            )
+        )
         . '</td>'
         . '<td class="nowrap" name="col_default">'
-        . PMA_getHtmlForColumnDefault(0, 3, 0, '', '', array())
+        . PMA\Template::get('columns_definitions/column_default')
+            ->render(array(
+                'columnNumber' => 0,
+                'ci' => 3,
+                'ci_offset' => 0,
+                'type_upper' => '',
+                'columnMeta' => array()
+            ))
         . '</td>'
         . '<td name="collation" class="nowrap">'
-        . PMA_getHtmlForColumnCollation(
-            0, 4, 0, array()
+        . PMA_generateCharsetDropdownBox(
+            PMA_CSDROPDOWN_COLLATION, 'field_collation[0]',
+            'field_0_4', null, false
         )
         . '</td>'
         . '<td class="nowrap" name="col_attribute">'
-        . PMA_getHtmlForColumnAttribute(0, 5, 0, array(), array(), false, null)
+        . PMA\Template::get('columns_definitions/column_attribute')
+            ->render(array(
+                'columnNumber' => 0,
+                'ci' => 5,
+                'ci_offset' => 0,
+                'extracted_columnspec' => array(),
+                'columnMeta' => array(),
+                'submit_attribute' => false,
+                'analyzed_sql' => null
+            ))
         . '</td>'
         . '<td class="nowrap" name="col_isNull">'
-        . PMA_getHtmlForColumnNull(0, 6, 0, array())
+        . PMA\Template::get('columns_definitions/column_null')
+            ->render(array(
+                'columnNumber' => 0,
+                'ci' => 6,
+                'ci_offset' => 0,
+                'columnMeta' => array()
+            ))
         . '</td>'
         . '<td class="nowrap" name="col_extra">'
-        . PMA_getHtmlForColumnExtra(0, 7, 0, array())
+        . PMA\Template::get('columns_definitions/column_extra')->render(
+            array(
+                'columnNumber' => 0,
+                'ci' => 7,
+                'ci_offset' => 0,
+                'columnMeta' => array()
+            )
+        )
         . '</td>'
         . ' <td>'
         . '<input id="add_column_save" type="submit" '
