+++ templates/columns_definitions/column_adjust_privileges.phtml
@@ -1,5 +1,5 @@
 <input name="field_adjust_privileges[<?php echo $columnNumber; ?>]"
-       id="field_'<?php echo $columnNumber; ?>_<?php echo ($ci - $ci_offset); ?>"
+       id="field_<?php echo $columnNumber; ?>_<?php echo ($ci - $ci_offset); ?>"
        checked="checked"
        type="checkbox"
        value="NULL"
