+++ templates/columns_definitions/column_comment.phtml
@@ -3,8 +3,8 @@
        name="field_comments[<?php echo $columnNumber; ?>]"
        size="12"
        maxlength="<?php echo PMA_MYSQL_INT_VERSION >= 50503 ? 1024 : 255; ?>"
-       value="<?php echo (isset($columnMeta['Field'])
+       value="<?php if (isset($columnMeta['Field'])
                             && is_array($comments_map)
-                            && isset($comments_map[$columnMeta['Field']]) ?
-                            htmlspecialchars($comments_map[$columnMeta['Field']]) : ''); ?>"
+                            && isset($comments_map[$columnMeta['Field']]))
+           echo htmlspecialchars($comments_map[$columnMeta['Field']]); ?>"
        class="textfield" />
