+++ templates/columns_definitions/column_default.phtml
@@ -1,3 +1,27 @@
+<?php
+// here we put 'NONE' as the default value of drop-down; otherwise
+// users would have problems if they forget to enter the default
+// value (example, for an INT)
+$default_options = array(
+    'NONE'              =>  _pgettext('for default', 'None'),
+    'USER_DEFINED'      =>  __('As defined:'),
+    'NULL'              => 'NULL',
+    'CURRENT_TIMESTAMP' => 'CURRENT_TIMESTAMP',
+);
+
+// for a TIMESTAMP, do not show the string "CURRENT_TIMESTAMP" as a default
+// value
+$defaultValue = '';
+if (isset($columnMeta['DefaultValue'])) {
+    $defaultValue = $columnMeta['DefaultValue'];
+}
+if ($type_upper == 'BIT') {
+    $defaultValue = PMA_Util::convertBitDefaultValue($columnMeta['DefaultValue']);
+} elseif ($type_upper == 'BINARY' || $type_upper == 'VARBINARY') {
+    $defaultValue = bin2hex($columnMeta['DefaultValue']);
+}
+?>
+
 <select name="field_default_type[<?php echo $columnNumber; ?>]"
         id="field_<?php echo $columnNumber; ?>_<?php echo($ci - $ci_offset); ?>"
         class="default_type">
@@ -16,12 +40,12 @@
               cols="15"
               class="textfield
               default_value">
-        <?php echo $default_value; ?>
+        <?php echo htmlspecialchars($defaultValue); ?>
     </textarea>
 <?php else: ?>
     <input type="text"
            name="field_default_value[<?php echo $columnNumber; ?>]"
            size="12"
-           value="<?php echo $default_value ?>"
+           value="<?php echo htmlspecialchars($defaultValue); ?>"
            class="textfield default_value" />
 <?php endif; ?>
