+++ templates/columns_definitions/column_name.phtml
@@ -1,0 +1,48 @@
+<?php
+$title = '';
+if (isset($columnMeta['column_status'])) {
+    if ($columnMeta['column_status']['isReferenced']) {
+        $title .= sprintf(
+            __('Referenced by %s.'),
+            implode(",", $columnMeta['column_status']['references'])
+        );
+    }
+    if ($columnMeta['column_status']['isForeignKey']) {
+        if (!empty($title)) {
+            $title .= "\n";
+        }
+        $title .=  __('Is a foreign key.');
+    }
+}
+if (empty($title)) {
+    $title = __('Column');
+}
+?>
+<input id="field_<?php echo $columnNumber; ?>_<?php echo ($ci - $ci_offset); ?>"
+       <?php if (isset($columnMeta['column_status'])
+           && !$columnMeta['column_status']['isEditable']): ?>
+           disabled="disabled"
+       <?php endif; ?>
+       type="text"
+       name="field_name[<?php echo $columnNumber; ?>]"
+       maxlength="64"
+       class="textfield"
+       title="<?php echo $title; ?>"
+       size="10"
+       value="<?php if (isset($columnMeta['Field']))
+           echo htmlspecialchars($columnMeta['Field']); ?>" />
+
+<?php if (isset($cfgRelation['central_columnswork'])
+    && $cfgRelation['central_columnswork']
+    && !(isset($columnMeta['column_status'])
+        && !$columnMeta['column_status']['isEditable'])
+): ?>
+    <p style="font-size:80%;margin:5px 2px"
+       id="central_columns_<?php echo $columnNumber; ?>_<?php echo ($ci - $ci_offset); ?>">
+        <a data-maxrows="<?php echo $GLOBALS['cfg']['MaxRows']; ?>"
+           href="#"
+           class="central_columns_dialog">
+            <?php echo __('Pick from Central Columns'); ?>
+        </a>
+    </p>
+<?php endif; ?>
