+++ templates/columns_definitions/column_type.phtml
@@ -1,7 +1,7 @@
 <select class="column_type"
         name="field_type[<?php echo $columnNumber; ?>]"
-        id="<?php echo $select_id; ?>"
+        id="field_<?php echo $columnNumber; ?>_<?php echo ($ci - $ci_offset); ?>"
         <?php if (isset($columnMeta['column_status']) && !$columnMeta['column_status']['isEditable'])
             echo 'disabled="disabled"'; ?>>
-<?php echo PMA_Util::getSupportedDatatypes(true, $type_upper); ?>
+    <?php echo PMA_Util::getSupportedDatatypes(true, $type_upper); ?>
 </select>
