+++ templates/designer/schema_export.phtml
@@ -8,4 +8,4 @@
         <input type="hidden" name="page_number" value="<?php echo htmlspecialchars($page); ?>" />
         <?php echo PMA_pluginGetOptions('Schema', $export_list); ?>
     </fieldset>
-</form>';
+</form>
