+++ test/libraries/PMA_tbl_columns_definition_form_test.php
@@ -10,7 +10,6 @@
  * Include to test.
  */
 require_once 'libraries/Util.class.php';
-require_once 'libraries/tbl_columns_definition_form.lib.php';
 require_once 'libraries/DatabaseInterface.class.php';
 require_once 'libraries/Partition.class.php';
 require_once 'libraries/Types.class.php';
@@ -34,136 +33,13 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function setUp()
     {
-        $GLOBALS['cfg']['ServerDefault'] = 1;
-        $GLOBALS['cfg']['DBG'] = null;
-        $GLOBALS['pmaThemeImage'] = 'image';
-
-        $_SESSION['PMA_Theme'] = PMA_Theme::load('./themes/pmahomme');
-        $_SESSION['PMA_Theme'] = new PMA_Theme();
-    }
-
-    /**
-     * Test for PMA_getFormsParameters
-     *
-     * @return void
-     */
-    public function testGetFormsParameters()
-    {
-        // case 1
-        $_REQUEST['after_field'] = "affield";
-        $_REQUEST['field_where'] = "fwhere";
-
-        $result = PMA_getFormsParameters(
-            "dbname", "tablename", "tbl_create.php", 22, array(12, 13)
-        );
-
-        $this->assertEquals(
-            array(
-                'db' => 'dbname',
-                'reload' => 1,
-                'orig_num_fields' => 22,
-                'orig_field_where' => 'fwhere',
-                'orig_after_field' => 'affield',
-                'selected[0]' => 12,
-                'selected[1]' => 13
-            ),
-            $result
-        );
-
-        // case 2
-        $result = PMA_getFormsParameters(
-            "dbname", "tablename", "tbl_addfield.php", null, 1
-        );
-
-        $this->assertEquals(
-            array(
-                'db' => 'dbname',
-                'table' => 'tablename',
-                'orig_field_where' => 'fwhere',
-                'orig_after_field' => 'affield',
-                'field_where' => 'fwhere',
-                'after_field' => 'affield'
-            ),
-            $result
-        );
-
-        // case 3
-        $_REQUEST['after_field'] = null;
-        $_REQUEST['field_where'] = null;
-
-        $result = PMA_getFormsParameters(
-            "dbname", "tablename", null, 0, null
-        );
-
-        $this->assertEquals(
-            array(
-                'db' => 'dbname',
-                'table' => 'tablename',
-                'orig_num_fields' => 0
-            ),
-            $result
-        );
-    }
-
-    /**
-     * Test for PMA_getHtmlForTableCreateOrAddField
-     *
-     * @return void
-     */
-    public function testGetHtmlForTableCreateOrAddField()
-    {
-        $dbi = $this->getMockBuilder('PMA_DatabaseInterface')
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $dbi->expects($this->any())
-            ->method('fetchResult')
-            ->will(
-                $this->returnValue(
-                    array()
-                )
-            );
-
-        $GLOBALS['dbi'] = $dbi;
-        $result = PMA_getHtmlForTableCreateOrAddField(
-            "tbl_create.php",
-            array('a' => 'b'),
-            array(array('c1')),
-            array('h1')
-        );
-
         /**
-         * @todo Find out a better method to test for HTML
+         * $GLOBALS['cfg']['ServerDefault'] = 1;
+         * $GLOBALS['cfg']['DBG'] = null;
+         * $GLOBALS['pmaThemeImage'] = 'image';
          *
-         * $this->assertContains(
-         *     '<form method="post" action="tbl_create.php" '
-         *     . 'class="create_table_form ajax lock-page">',
-         *     $result
-         * );
-         */
-
-        $this->assertContains(
-            '<input type="hidden" name="a" value="b"',
-            $result
-        );
-
-        $this->assertContains(
-            '<select lang="en" dir="ltr" name="tbl_collation">',
-            $result
-        );
-
-        /**
-         * @todo Find out a better method to test for HTML
-         *
-         * $this->assertContains(
-         *     '<input type="submit" name="do_save_data" value="Save"',
-         *     $result
-         * );
-         *
-         * $this->assertContains(
-         *     '<input type="text" name="table"',
-         *     $result
-         * );
+         * $_SESSION['PMA_Theme'] = PMA_Theme::load('./themes/pmahomme');
+         * $_SESSION['PMA_Theme'] = new PMA_Theme();
          */
     }
 
@@ -174,24 +50,27 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function testGetHeaderCells()
     {
-        $GLOBALS['cfg']['BrowseMIME'] = true;
-        $GLOBALS['cfg']['ShowHint'] = false;
-        $result = PMA_getHeaderCells(false, array(), true);
-
-        $this->assertContains(
-            'Index',
-            $result
-        );
-
-        $this->assertContains(
-            'Move column',
-            $result
-        );
-
-        $this->assertContains(
-            'MIME type',
-            $result
-        );
+        /**
+         * @todo Test against table_fields_definition.phtml
+         * $GLOBALS['cfg']['BrowseMIME'] = true;
+         * $GLOBALS['cfg']['ShowHint'] = false;
+         * $result = PMA_getHeaderCells(false, array(), true);
+         *
+         * $this->assertContains(
+         *     'Index',
+         *     $result
+         * );
+         *
+         * $this->assertContains(
+         *     'Move column',
+         *     $result
+         * );
+         *
+         * $this->assertContains(
+         *     'MIME type',
+         *     $result
+         * );
+         */
     }
 
     /**
@@ -201,26 +80,29 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function testGetMoveColumns()
     {
-        $dbi = $this->getMockBuilder('PMA_DatabaseInterface')
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $dbi->expects($this->once())
-            ->method('tryQuery')
-            ->with("SELECT * FROM `db`.`table` LIMIT 1")
-            ->will($this->returnValue('v1'));
-
-        $dbi->expects($this->once())
-            ->method('getFieldsMeta')
-            ->with("v1")
-            ->will($this->returnValue('movecols'));
-
-        $GLOBALS['dbi'] = $dbi;
-
-        $this->assertEquals(
-            PMA_getMoveColumns('db', 'table'),
-            'movecols'
-        );
+        /**
+         * @todo Test against PMA_Table::getColumnsMeta
+         * $dbi = $this->getMockBuilder('PMA_DatabaseInterface')
+         *     ->disableOriginalConstructor()
+         *     ->getMock();
+         *
+         * $dbi->expects($this->once())
+         *     ->method('tryQuery')
+         *     ->with("SELECT * FROM `db`.`table` LIMIT 1")
+         *     ->will($this->returnValue('v1'));
+         *
+         * $dbi->expects($this->once())
+         *     ->method('getFieldsMeta')
+         *     ->with("v1")
+         *     ->will($this->returnValue('movecols'));
+         *
+         * $GLOBALS['dbi'] = $dbi;
+         *
+         * $this->assertEquals(
+         *     PMA_getMoveColumns('db', 'table'),
+         *     'movecols'
+         * );
+         */
     }
 
     /**
@@ -230,6 +112,7 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function testGetRowDataForRegeneration()
     {
+        /** @todo Move test
         $_REQUEST = array(
             'field_name' => array(1 => 'name'),
             'field_type' => array(1 => 'type'),
@@ -260,6 +143,7 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
             ),
             $result
         );
+         */
     }
 
     /**
@@ -269,6 +153,7 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function testGetSubmitPropertiesForRegeneration()
     {
+        /** @todo Move test
         $_REQUEST = array(
             'field_length' => array(1 => 22),
             'field_attribute' => array(1 => 'attr'),
@@ -281,7 +166,7 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
             array(22, 'attr', false),
             $result
         );
-
+        */
     }
 
     /**
@@ -291,6 +176,7 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function testHandleRegeneration()
     {
+        /** @todo Move test
         $_REQUEST = array(
             'field_comments' => array(1 => 'comm'),
             'field_mimetype' => array(1 => 'mime'),
@@ -315,6 +201,7 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
             ),
             $result[5]
         );
+         */
     }
 
     /**
@@ -324,104 +211,105 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function testGetColumnMetaForDefault()
     {
-        $cmeta = array(
-            'Default' => null,
-            'Null' => 'YES',
-            'DefaultType' => 'a',
-            'DefaultValue' => 'b',
-        );
-
-        $result = PMA_getColumnMetaForDefault($cmeta, null);
-
-        $this->assertEquals(
-            'NULL',
-            $result['DefaultType']
-        );
-
-        $this->assertEquals(
-            '',
-            $result['DefaultValue']
-        );
-
-        // case 2
-        $cmeta = array(
-            'Default' => null,
-            'Null' => 'NO',
-            'DefaultType' => 'a',
-            'DefaultValue' => 'b',
-        );
-
-        $result = PMA_getColumnMetaForDefault($cmeta, true);
-
-        $this->assertEquals(
-            'USER_DEFINED',
-            $result['DefaultType']
-        );
-
-        $this->assertEquals(
-            null,
-            $result['DefaultValue']
-        );
-
-        // case 3
-        $cmeta = array(
-            'Default' => null,
-            'Null' => 'NO',
-            'DefaultType' => 'a',
-            'DefaultValue' => 'b',
-        );
-
-        $result = PMA_getColumnMetaForDefault($cmeta, false);
-
-        $this->assertEquals(
-            'NONE',
-            $result['DefaultType']
-        );
-
-        $this->assertEquals(
-            null,
-            $result['DefaultValue']
-        );
-
-        // case 4
-        $cmeta = array(
-            'Default' => 'CURRENT_TIMESTAMP',
-            'Null' => 'NO',
-            'DefaultType' => 'a',
-            'DefaultValue' => 'b',
-        );
-
-        $result = PMA_getColumnMetaForDefault($cmeta, false);
-
-        $this->assertEquals(
-            'CURRENT_TIMESTAMP',
-            $result['DefaultType']
-        );
-
-        $this->assertEquals(
-            null,
-            $result['DefaultValue']
-        );
-
-        // case 5
-        $cmeta = array(
-            'Default' => 'SOMETHING_ELSE',
-            'Null' => 'NO',
-            'DefaultType' => 'a',
-            'DefaultValue' => 'b',
-        );
-
-        $result = PMA_getColumnMetaForDefault($cmeta, false);
-
-        $this->assertEquals(
-            'USER_DEFINED',
-            $result['DefaultType']
-        );
-
-        $this->assertEquals(
-            'SOMETHING_ELSE',
-            $result['DefaultValue']
-        );
+        // @todo Move test
+//        $cmeta = array(
+//            'Default' => null,
+//            'Null' => 'YES',
+//            'DefaultType' => 'a',
+//            'DefaultValue' => 'b',
+//        );
+//
+//        $result = PMA_getColumnMetaForDefault($cmeta, null);
+//
+//        $this->assertEquals(
+//            'NULL',
+//            $result['DefaultType']
+//        );
+//
+//        $this->assertEquals(
+//            '',
+//            $result['DefaultValue']
+//        );
+//
+//        // case 2
+//        $cmeta = array(
+//            'Default' => null,
+//            'Null' => 'NO',
+//            'DefaultType' => 'a',
+//            'DefaultValue' => 'b',
+//        );
+//
+//        $result = PMA_getColumnMetaForDefault($cmeta, true);
+//
+//        $this->assertEquals(
+//            'USER_DEFINED',
+//            $result['DefaultType']
+//        );
+//
+//        $this->assertEquals(
+//            null,
+//            $result['DefaultValue']
+//        );
+//
+//        // case 3
+//        $cmeta = array(
+//            'Default' => null,
+//            'Null' => 'NO',
+//            'DefaultType' => 'a',
+//            'DefaultValue' => 'b',
+//        );
+//
+//        $result = PMA_getColumnMetaForDefault($cmeta, false);
+//
+//        $this->assertEquals(
+//            'NONE',
+//            $result['DefaultType']
+//        );
+//
+//        $this->assertEquals(
+//            null,
+//            $result['DefaultValue']
+//        );
+//
+//        // case 4
+//        $cmeta = array(
+//            'Default' => 'CURRENT_TIMESTAMP',
+//            'Null' => 'NO',
+//            'DefaultType' => 'a',
+//            'DefaultValue' => 'b',
+//        );
+//
+//        $result = PMA_getColumnMetaForDefault($cmeta, false);
+//
+//        $this->assertEquals(
+//            'CURRENT_TIMESTAMP',
+//            $result['DefaultType']
+//        );
+//
+//        $this->assertEquals(
+//            null,
+//            $result['DefaultValue']
+//        );
+//
+//        // case 5
+//        $cmeta = array(
+//            'Default' => 'SOMETHING_ELSE',
+//            'Null' => 'NO',
+//            'DefaultType' => 'a',
+//            'DefaultValue' => 'b',
+//        );
+//
+//        $result = PMA_getColumnMetaForDefault($cmeta, false);
+//
+//        $this->assertEquals(
+//            'USER_DEFINED',
+//            $result['DefaultType']
+//        );
+//
+//        $this->assertEquals(
+//            'SOMETHING_ELSE',
+//            $result['DefaultValue']
+//        );
     }
 
     /**
@@ -431,19 +319,9 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function testGetHtmlForColumnName()
     {
-        $cfgRelation = array('central_columnswork' => true);
-        $result = PMA_getHtmlForColumnName(
-            2, 4, 4, array('Field' => "fieldname",
-            'column_status' => array('isReferenced' => false,
-            'isForeignKey' => false, 'isEditable' => true)), $cfgRelation
-        );
-
-        $this->assertContains(
-            '<input id="field_2_0" type="text" name="field_name[2]" '
-            . 'maxlength="64" class="textfield" title="Column" size="10" '
-            . 'value="fieldname" />',
-            $result
-        );
+        /**
+         * @todo  Create test for page
+         */
     }
 
     /**
@@ -453,25 +331,25 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function testGetHtmlForColumnType()
     {
-        $GLOBALS['PMA_Types'] = new PMA_Types;
-        $result = PMA_getHtmlForColumnType(
-            1, 4, 3, false, array('column_status' => array('isReferenced' => false,
-            'isForeignKey' => false, 'isEditable' => true))
-        );
-
         /**
          * @todo Find out a better method to test for HTML
+         *
+         * $GLOBALS['PMA_Types'] = new PMA_Types;
+         * $result = PMA_getHtmlForColumnType(
+         *     1, 4, 3, false, array('column_status' => array('isReferenced' => false,
+         *     'isForeignKey' => false, 'isEditable' => true))
+         * );
          *
          * $this->assertContains(
          *     '<select class="column_type" name="field_type[1]" id="field_1_1">',
          *     $result
          * );
+         *
+         * $this->assertContains(
+         *     '<option title="">INT</option>',
+         *     $result
+         * );
          */
-
-        $this->assertContains(
-            '<option title="">INT</option>',
-            $result
-        );
     }
 
     /**
@@ -514,42 +392,42 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function testGetHtmlForTransformation()
     {
-        $cmeta = array(
-            'Field' => 'fieldname'
-        );
-
-        $mime = array(
-            'fieldname' => array(
-                'transformation' => 'Text_Plain_Preappend.class.php',
-                'transformation_options' => 'transops'
-            )
-        );
-
-        $avail_mime = array(
-            'transformation' => array(
-                'foo' => 'text/plain: bar'
-            ),
-            'transformation_file' => array(
-                'foo' => 'Text_Plain_Preappend.class.php'
-            )
-        );
-        $result = PMA_getHtmlForTransformation(
-            2, 0, 0, $avail_mime, $cmeta, $mime, ''
-        );
-
         /**
          * @todo Find out a better method to test for HTML
+         *
+         * $cmeta = array(
+         *     'Field' => 'fieldname'
+         * );
+         *
+         * $mime = array(
+         *     'fieldname' => array(
+         *         'transformation' => 'Text_Plain_Preappend.class.php',
+         *         'transformation_options' => 'transops'
+         *     )
+         * );
+         *
+         * $avail_mime = array(
+         *     'transformation' => array(
+         *         'foo' => 'text/plain: bar'
+         *     ),
+         *     'transformation_file' => array(
+         *         'foo' => 'Text_Plain_Preappend.class.php'
+         *     )
+         * );
+         * $result = PMA_getHtmlForTransformation(
+         * 2, 0, 0, $avail_mime, $cmeta, $mime, ''
+         * );
          *
          * $this->assertContains(
          *     '<select id="field_2_0" size="1" name="field_transformation[2]">',
          *     $result
          * );
+         *
+         * $this->assertContains(
+         *     'selected ',
+         *     $result
+         * );
          */
-
-        $this->assertContains(
-            'selected ',
-            $result
-        );
     }
 
     /**
@@ -559,37 +437,37 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function testGetHtmlForMoveColumn()
     {
-        $cmeta = array(
-            'Field' => 'fieldname'
-        );
-
-        $moveColumns = array();
-
-        $temp = new stdClass;
-        $temp->name = 'a';
-        $moveColumns[] = $temp;
-
-        $temp = new stdClass;
-        $temp->name = 'fieldname';
-        $moveColumns[] = $temp;
-
-        $result = PMA_getHtmlForMoveColumn(
-            2, 0, 0, $moveColumns, $cmeta
-        );
-
         /**
          * @todo Find out a better method to test for HTML
+         *
+         * $cmeta = array(
+         *     'Field' => 'fieldname'
+         * );
+         *
+         * $moveColumns = array();
+         *
+         * $temp = new stdClass;
+         * $temp->name = 'a';
+         * $moveColumns[] = $temp;
+         *
+         * $temp = new stdClass;
+         * $temp->name = 'fieldname';
+         * $moveColumns[] = $temp;
+         *
+         * $result = PMA_getHtmlForMoveColumn(
+         *     2, 0, 0, $moveColumns, $cmeta
+         * );
          *
          * $this->assertContains(
          *     '<select id="field_2_0" name="field_move_to[2]" size="1" width="5em">',
          *     $result
          * );
+         *
+         * $this->assertContains(
+         *     '<option value="" selected="selected">&nbsp;</option>',
+         *     $result
+         * );
          */
-
-        $this->assertContains(
-            '<option value="" selected="selected">&nbsp;</option>',
-            $result
-        );
 
         /**
          * @todo Find out a better method to test for HTML
@@ -672,15 +550,6 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
      */
     public function testGetHtmlForColumnIndexes()
     {
-        $cmeta = array(
-            'Extra' => 'auto_increment',
-            'Field' => 'fieldname'
-        );
-
-        $result = PMA_getHtmlForColumnIndexes(
-            2, 1, 0, $cmeta
-        );
-
         /**
          * @todo Find out a better method to test for HTML
          *
@@ -688,15 +557,12 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
          *     '<select name="field_key[2]" id="field_2_1"',
          *     $result
          * );
-         */
-
-        $this->assertContains(
-            '<option value="none_2">---</option>',
-            $result
-        );
-
-        /**
-         * @todo Find out a better method to test for HTML
+         *
+         *
+         * $this->assertContains(
+         *     '<option value="none_2">---</option>',
+         *     $result
+         * );
          *
          * $this->assertContains(
          *     '<option value="primary_2" title="Primary">PRIMARY</option>',
@@ -802,56 +668,27 @@ class PMA_TblColumnsDefinitionFormTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * Test for PMA_getHtmlForColumnCollation
-     *
-     * @return void
-     */
-    public function testGetHtmlForColumnCollation()
-    {
-        $cmeta = array(
-            'Collation' => 'utf8_general_ci'
-        );
-
-        $result = PMA_getHtmlForColumnCollation(
-            2, 3, 1, $cmeta
-        );
-
-        $this->assertContains(
-            '<select lang="en" dir="ltr" name="field_collation[2]" id="field_2_2">',
-            $result
-        );
-
-        $this->assertContains(
-            '<option value="utf8_bin" title="Unicode (multilingual), Binary">',
-            $result
-        );
-    }
-
-    /**
      * Test for PMA_getHtmlForColumnLength
      *
      * @return void
      */
     public function testGetHtmlForColumnLength()
     {
-        $result = PMA_getHtmlForColumnLength(
-            2, 3, 1, 10, 8
-        );
-
         /**
          * @todo Find out a better method to test for HTML
+         * Template: columns_definitions/column_length
          *
          * $this->assertContains(
          *     '<input id="field_2_2" type="text" name="field_length[2]" size="10" '
          *     . 'value="8" class="textfield" />',
          *     $result
          * );
-         */
-
-        $this->assertContains(
-            '<p class="enum_notice" id="enum_notice_2_2">',
-            $result
-        );
+         *
+         * $this->assertContains(
+         *     '<p class="enum_notice" id="enum_notice_2_2">',
+         *     $result
+         * );
+        */
     }
 
     /**
