+++ test/libraries/PMA_tbl_relation_test.php
@@ -9,7 +9,6 @@
 /*
  * Include to test.
  */
-require_once 'libraries/tbl_relation.lib.php';
 require_once 'libraries/Util.class.php';
 require_once 'libraries/url_generating.lib.php';
 require_once 'libraries/database_interface.inc.php';
@@ -49,123 +48,6 @@ class PMA_TblRelationTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * Tests for PMA_generateRelationalDropdown() method.
-     *
-     * @return void
-     * @test
-     */
-    public function testGenerateRelationalDropdown()
-    {
-        // test for start tag
-        $this->assertStringStartsWith(
-            '<select',
-            PMA_generateRelationalDropdown('name')
-        );
-
-        // test for end tag
-        $this->assertStringEndsWith(
-            '</select>',
-            PMA_generateRelationalDropdown('name')
-        );
-
-        // test for name
-        $this->assertStringStartsWith(
-            '<select name="name"',
-            PMA_generateRelationalDropdown('name')
-        );
-
-        // test for title
-        $this->assertStringStartsWith(
-            '<select name="name" title="title"',
-            PMA_generateRelationalDropdown('name', array(), false, 'title')
-        );
-
-        $values = array('value1', '<alue2', 'value3');
-        // test for empty option
-        $this->assertContains(
-            '<option value=""></option>',
-            PMA_generateRelationalDropdown('name', $values)
-        );
-
-        /**
-         * @todo Find out a better method to test for HTML
-         *
-         * // test for options and escaping
-         * $this->assertContains(
-         *     '<option value="&lt;alue2">&lt;alue2</option>',
-         *     PMA_generateRelationalDropdown('name', $values)
-         * );
-         *
-         * // test for selected option
-         * $this->assertContains(
-         *     '<option value="value1" selected="selected">value1</option>',
-         *     PMA_generateRelationalDropdown('name', $values, 'value1')
-         * );
-         *
-         * // test for selected value not found in values array and its escaping
-         * $this->assertContains(
-         *     '<option value="valu&lt;4" selected="selected">valu&lt;4'
-         *     . '</option></select>',
-         *     PMA_generateRelationalDropdown('name', $values, 'valu<4')
-         * );
-         */
-    }
-
-    /**
-     * Tests for PMA_generateDropdown() method.
-     *
-     * @return void
-     * @test
-     */
-    public function testPMAGenerateDropdown()
-    {
-        $dropdown_question = "dropdown_question";
-        $select_name = "select_name";
-        $choices = array("choice1", "choice2");
-        $selected_value = "";
-
-        $html_output = PMA_generateDropdown(
-            $dropdown_question, $select_name, $choices, $selected_value
-        );
-
-        $this->assertContains(
-            htmlspecialchars($dropdown_question),
-            $html_output
-        );
-
-        $this->assertContains(
-            htmlspecialchars($select_name),
-            $html_output
-        );
-
-        $this->assertContains(
-            htmlspecialchars("choice1"),
-            $html_output
-        );
-
-        $this->assertContains(
-            htmlspecialchars("choice2"),
-            $html_output
-        );
-    }
-
-    /**
-     * Tests for PMA_backquoteSplit() method.
-     *
-     * @return void
-     * @test
-     */
-    public function testPMABackquoteSplit()
-    {
-        $text = "test `PMA` Back `quote` Split";
-
-        $this->assertEquals(
-            array('`PMA`', '`quote`'),
-            PMA_backquoteSplit($text)
-        );
-    }
-
-    /**
      * Tests for PMA_getSQLToCreateForeignKey() method.
      *
      * @return void
@@ -173,7 +55,8 @@ class PMA_TblRelationTest extends PHPUnit_Framework_TestCase
      */
     public function testPMAGetSQLToCreateForeignKey()
     {
-        $table = "PMA_table";
+        // @todo Move this test to PMA_Table_test
+/*        $table = "PMA_table";
         $field = array("PMA_field1", "PMA_field2");
         $foreignDb = "foreignDb";
         $foreignTable = "foreignTable";
@@ -188,24 +71,7 @@ class PMA_TblRelationTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(
             $sql_excepted,
             $sql
-        );
-    }
-
-    /**
-     * Tests for PMA_getSQLToDropForeignKey() method.
-     *
-     * @return void
-     * @test
-     */
-    public function testPMAGetSQLToDropForeignKey()
-    {
-        $table = "pma_table";
-        $fk = "pma_fk";
-
-        $this->assertEquals(
-            "ALTER TABLE `pma_table` DROP FOREIGN KEY `pma_fk`;",
-            PMA_getSQLToDropForeignKey($table, $fk)
-        );
+        );*/
     }
 
     /**
@@ -216,83 +82,19 @@ class PMA_TblRelationTest extends PHPUnit_Framework_TestCase
      */
     public function testPMAGetHtmlForCommonForm()
     {
-        $db = "pma_db";
-        $table = "pma_table";
-        $columns = array(
-            array("Field" => "Field1")
-        );
-        $cfgRelation = array(
-            'displaywork' => true,
-            'relwork' => true,
-            'displaywork' => true,
-        );
-        $tbl_storage_engine = "InnoDB";
-        $existrel =  array();
-        $existrel_foreign =  array();
-        $options_array =  array();
-
-        $save_row =  array();
-        foreach ($columns as $row) {
-            $save_row[] = $row;
-        }
-
-        $html = PMA_getHtmlForCommonForm(
-            $db, $table, $columns, $cfgRelation,
-            $tbl_storage_engine, $existrel, $existrel_foreign, $options_array
-        );
-
-        $this->assertContains(
-            PMA_URL_getHiddenInputs($db, $table),
-            $html
-        );
-
-        $this->assertContains(
-            __('Column'),
-            $html
-        );
-
-        $this->assertContains(
-            __('Internal relation'),
-            $html
-        );
-
-        $this->assertContains(
-            __('Choose column to display:'),
-            $html
-        );
-
-        /* @todo Find out a better method to test for HTML
-         * //case 3: PMA_getHtmlForInternalRelationRow
-         *$row = PMA_getHtmlForInternalRelationRow(
-         *   $save_row, 0, true,
-         *   $existrel, $db
-         *);
-         *$this->assertContains(
-         *   $row,
-         *   $html
-         *);
-         */
-
-        //case 4: PMA_getHtmlForForeignKeyRow
-        $row = PMA_getHtmlForForeignKeyRow(
-            array(), true, $columns, 0,
-            $options_array, $tbl_storage_engine, $db
-        );
-        $this->assertContains(
-            $row,
-            $html
-        );
-
+        // @todo Find out a better method to test for HTML
     }
 
     /**
      * Tests for PMA_getQueryForDisplayUpdate() method.
+     * @todo Move this test to PMA_Table_test
      *
      * @return void
      * @test
      */
     public function testPMAGetQueryForDisplayUpdate()
     {
+    /*
         $disp = true;
         $display_field = '';
         $db = "pma_db";
@@ -353,7 +155,7 @@ class PMA_TblRelationTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(
             $query_expect,
             $query
-        );
+        );*/
     }
 }
